+++ src/resolvers.ts
@@ -1,4 +1,4 @@
-import { AppDataSource } from "./dataSource";
+import { AppDataSource } from "./data-source";
 import { User } from "./entities/user";
 
 const UserRepository = AppDataSource.getRepository(User);
